fix: redundant type cast

+++ opc-ua-sdk/codec-json/src/test/java/org/eclipse/milo/sdk/core/types/json/JsonConversionsTest.java
@@ -157,7 +157,7 @@ class JsonConversionsTest {
     assertEquals(new JsonPrimitive(input), jsonValue);
 
     float opcValue = JsonConversions.toFloat(jsonValue);
-    assertEquals((float) input, opcValue);
+    assertEquals(input, opcValue);
   }
 
   @ParameterizedTest
